+++ CHANGELOG.md
@@ -24,6 +24,7 @@
 - Add dark mode of Chat component [#125](https://github.com/mishka-group/mishka_chelekom/pull/125)
 - Add dark mode of Modal component [#126](https://github.com/mishka-group/mishka_chelekom/pull/126)
 - Add dark mode of Keyboard component [#127](https://github.com/mishka-group/mishka_chelekom/pull/127)
+- Add dark mode of Jumbotron component [#128](https://github.com/mishka-group/mishka_chelekom/pull/128)
 
 ### Bugs:
 - Fix un-CSP progress mounted [#72](https://github.com/mishka-group/mishka_chelekom/pull/72)